Create Steps to identify a palindrome
To find if a sentence is palindrome, compare each character from left and right. If they are equal, compare until left and right of string are equal or right becomes less than left.

For example. the word "RACECAR" is a palindrome, because when you write it backwards, it is "RACECAR" again. 
The word "DEFIED" is not a palindrome, because when you write it backwards, it is "DEIFED". 
The word "MADAM" is a palindrome because when you spell it backwards, it is "MADAM" agaim.
Palindrome words: Radar, Level, did, noon, mom, rotator, deed and etc.